Forms: extract WidgetEditorReadonlyView into its own component
Move the widget editor readonly view into a separate component so that
useBlockEditingMode('contentOnly') is only called when the component
renders, avoiding setting 'default' mode in non-widget contexts.
