+++ projects/packages/forms/src/blocks/contact-form/components/widget-editor-readonly-view.scss
@@ -1,0 +1,11 @@
+.wp-block-jetpack-contact-form {
+
+	.jetpack-contact-form-widget-readonly-notice {
+		margin: 8px 0 4px;
+
+		.components-notice__action.components-button {
+			margin-left: 0;
+			margin-top: 8px;
+		}
+	}
+}
