+++ projects/packages/forms/src/blocks/contact-form/components/widget-editor-readonly-view.tsx
@@ -1,0 +1,56 @@
+import { useBlockEditingMode } from '@wordpress/block-editor';
+import { Disabled, Notice } from '@wordpress/components';
+import { __ } from '@wordpress/i18n';
+import { navigateToForm } from './convert-form-toolbar.tsx';
+import ContactFormSkeletonLoader from './jetpack-contact-form-skeleton-loader.js';
+import './widget-editor-readonly-view.scss';
+
+export default function WidgetEditorReadonlyView( {
+	blockProps,
+	innerBlocksProps,
+	isResolvingSyncedForm,
+	formRef,
+	flushPendingSave,
+}: {
+	blockProps: Record< string, unknown >;
+	innerBlocksProps: Record< string, unknown >;
+	isResolvingSyncedForm: boolean;
+	formRef: number;
+	flushPendingSave: () => void;
+} ) {
+	useBlockEditingMode( 'contentOnly' );
+
+	const handleEditForm = () => {
+		flushPendingSave();
+		navigateToForm( formRef, 'widget' );
+	};
+
+	return (
+		<div { ...blockProps }>
+			<Notice
+				className="jetpack-contact-form-widget-readonly-notice"
+				status="info"
+				isDismissible={ false }
+				actions={ [
+					{
+						label: __( 'Edit Form', 'jetpack-forms' ),
+						onClick: handleEditForm,
+						variant: 'primary',
+					},
+				] }
+			>
+				{ __(
+					'Forms are edited in the Form Editor. Changes will sync back to this widget.',
+					'jetpack-forms'
+				) }
+			</Notice>
+			{ isResolvingSyncedForm ? (
+				<ContactFormSkeletonLoader />
+			) : (
+				<Disabled>
+					<div { ...innerBlocksProps } />
+				</Disabled>
+			) }
+		</div>
+	);
+}
