+++ projects/packages/forms/src/blocks/contact-form/edit.tsx
@@ -13,11 +13,9 @@ import {
 	store as blockEditorStore,
 	BlockControls,
 	BlockContextProvider,
-	useBlockEditingMode,
 } from '@wordpress/block-editor';
 import { createBlock } from '@wordpress/blocks';
 import {
-	Disabled,
 	ExternalLink,
 	Notice,
 	PanelBody,
@@ -51,12 +49,13 @@ import useFormSteps from '../shared/hooks/use-form-steps.js';
 import { SyncedAttributeProvider } from '../shared/hooks/use-synced-attributes.js';
 import { CORE_BLOCKS, FORM_POST_TYPE } from '../shared/util/constants.js';
 import { childBlocks } from './child-blocks.js';
-import { ConvertFormToolbar, navigateToForm } from './components/convert-form-toolbar.tsx';
+import { ConvertFormToolbar } from './components/convert-form-toolbar.tsx';
 import FormStatusNotice from './components/form-status-notice.tsx';
 import { ContactFormPlaceholder } from './components/jetpack-contact-form-placeholder.js';
 import ContactFormSkeletonLoader from './components/jetpack-contact-form-skeleton-loader.js';
 import NotificationsSettings from './components/notifications-settings.js';
 import WebhooksSettings from './components/webhooks-settings.js';
+import WidgetEditorReadonlyView from './components/widget-editor-readonly-view.tsx';
 import { useSyncedFormAutoSave } from './hooks/use-synced-form-auto-save.ts';
 import { useSyncedFormLoader } from './hooks/use-synced-form-loader.ts';
 import { useSyncedForm } from './hooks/use-synced-form.ts';
@@ -215,8 +214,6 @@ function JetpackContactFormEdit( {
 
 	// Check if we're in widget editor with a synced form (ref)
 	const isWidgetEditorWithRef = !! ref && getEditorContext() === 'widget';
-
-	useBlockEditingMode( isWidgetEditorWithRef ? 'contentOnly' : 'default' );
 
 	// Load synced form data from the jetpack_form post type
 	const {
@@ -1022,38 +1019,14 @@ function JetpackContactFormEdit( {
 	}
 	// In widget editor, synced forms (with ref) are not editable
 	else if ( isWidgetEditorWithRef ) {
-		const handleEditForm = () => {
-			flushPendingSave();
-			navigateToForm( ref as number, 'widget' );
-		};
-
 		return (
-			<div { ...blockProps }>
-				<Notice
-					className="jetpack-contact-form-widget-readonly-notice"
-					status="info"
-					isDismissible={ false }
-					actions={ [
-						{
-							label: __( 'Edit Form', 'jetpack-forms' ),
-							onClick: handleEditForm,
-							variant: 'primary',
-						},
-					] }
-				>
-					{ __(
-						'Forms are edited in the Form Editor. Changes will sync back to this widget.',
-						'jetpack-forms'
-					) }
-				</Notice>
-				{ isResolvingSyncedForm ? (
-					<ContactFormSkeletonLoader />
-				) : (
-					<Disabled>
-						<div { ...innerBlocksProps } />
-					</Disabled>
-				) }
-			</div>
+			<WidgetEditorReadonlyView
+				blockProps={ blockProps }
+				innerBlocksProps={ innerBlocksProps }
+				isResolvingSyncedForm={ isResolvingSyncedForm }
+				formRef={ ref as number }
+				flushPendingSave={ flushPendingSave }
+			/>
 		);
 	} else if ( ! isModuleActive ) {
 		if ( isLoadingModules ) {
