+++ projects/packages/forms/src/blocks/contact-form/editor.scss
@@ -210,15 +210,6 @@
 		}
 	}
 
-	/* Fixes notices styling when Gutenberg is not there */
-	.jetpack-contact-form-widget-readonly-notice {
-		margin: 8px 0 4px;
-
-		.components-notice__action.components-button {
-			margin-left: 0;
-			margin-top: 8px;
-		}
-	}
 }
 
 .jetpack-contact-form .components-placeholder {
